+++ src/declaration/kubernetes-address.ts
@@ -1,13 +1,14 @@
+import type { DeclarationConstraintOptions } from "../constraint.js";
 import { createHostnameConstraint } from "../constraint/hostname.js";
 import { createNetworkPortNumberConstraint } from "../constraint/network-port-number.js";
 import {
   Declaration,
   Options as DeclarationOptions,
-  Value,
   defaultFromOptions,
   type ExactOptions,
+  type Value,
 } from "../declaration.js";
-import { registerVariable } from "../environment.js";
+import { registerVariable, registerVariableComposite } from "../environment.js";
 import { SpecError, normalize } from "../error.js";
 import { resolveExamples, type Example } from "../example.js";
 import { Maybe, map, resolve } from "../maybe.js";
@@ -24,10 +25,11 @@ export type KubernetesAddress = {
   readonly port: number;
 };
 
-export type Options = DeclarationOptions<KubernetesAddress> & {
-  readonly examples?: KubernetesAddressExamples;
-  readonly portName?: string;
-};
+export type Options = DeclarationOptions<KubernetesAddress> &
+  DeclarationConstraintOptions<KubernetesAddress> & {
+    readonly examples?: KubernetesAddressExamples;
+    readonly portName?: string;
+  };
 
 type KubernetesAddressExamples = {
   readonly host?: Example<string>[];
@@ -39,6 +41,7 @@ export function kubernetesAddress<O extends Options>(
   options: ExactOptions<O, Options> = {} as ExactOptions<O, Options>,
 ): Declaration<KubernetesAddress, O> {
   const {
+    constraints: customConstraints = [],
     examples: { host: hostExamples, port: portExamples } = {},
     isSensitive = false,
     portName,
@@ -47,21 +50,17 @@ export function kubernetesAddress<O extends Options>(
   const def = defaultFromOptions(options);
 
   const hostVar = registerHost(name, hostExamples, isSensitive, def);
-  const hName = hostVar.spec.name;
   const portVar = registerPort(name, portExamples, isSensitive, def, portName);
-  const pName = portVar.spec.name;
+
+  const composite = registerVariableComposite({
+    variables: { host: hostVar, port: portVar },
+    resolve: ({ host, port }) => ({ host, port }),
+    constraints: [...customConstraints],
+  });
 
   return {
     value() {
-      const host = resolve(hostVar.nativeValue());
-      const port = resolve(portVar.nativeValue());
-
-      if (host != null && port != null) return { host, port };
-
-      if (host != null) throw new PartiallyDefinedError(hName, pName);
-      if (port != null) throw new PartiallyDefinedError(pName, hName);
-
-      return undefined as Value<KubernetesAddress, O>;
+      return resolve(composite.value()) as Value<KubernetesAddress, O>;
     },
   };
 }
@@ -205,11 +204,5 @@ class InvalidPortNameError extends SpecError {
       `Kubernetes ${name} service address`,
       new Error(`port name (${quotedName}): ${cause.message}`),
     );
-  }
-}
-
-class PartiallyDefinedError extends Error {
-  constructor(def: string, undef: string) {
-    super(`${def} is defined but ${undef} is not, define both or neither`);
   }
 }