+++ src/validation.ts
@@ -1,21 +1,47 @@
 import { normalize } from "./error.js";
 import { Maybe } from "./maybe.js";
+import {
+  VariableCompositeError,
+  type VariableComposite,
+} from "./variable-composite.js";
 import { Value, Variable } from "./variable.js";
 
-export function validate(variables: Variable<unknown>[]): [boolean, Results] {
-  const results: Results = [];
+export function validate(
+  variables: Variable<unknown>[],
+  composites: VariableComposite<unknown, Record<string, Variable<unknown>>>[],
+): [boolean, Results] {
+  const resultMap = new Map<Variable<unknown>, Result>();
   let isValid = true;
 
   for (const variable of variables) {
     try {
-      results.push({
-        variable,
-        result: { maybe: variable.value() },
-      });
+      resultMap.set(variable, { maybe: variable.value() });
     } catch (error) {
       isValid = false;
-      results.push({ variable, result: { error: normalize(error) } });
+
+      resultMap.set(variable, { error: normalize(error) });
     }
+  }
+
+  for (const composite of composites) {
+    try {
+      composite.value();
+    } catch (compositeError) {
+      isValid = false;
+
+      if (compositeError instanceof VariableCompositeError) {
+        for (const [variable, error] of compositeError.blame.entries()) {
+          resultMap.set(variable, { error });
+        }
+      }
+    }
+  }
+
+  const results: Results = [];
+
+  for (const variable of variables) {
+    const result = resultMap.get(variable);
+    if (result) results.push({ variable, result });
   }
 
   return [isValid, results];